requirements.txt
Django==4.0.1
pycryptodome==3.12.0
PyJWT==2.3.0
